psunote/requirements.txt
flask
flask-wtf
flask-sqlalchemy
wtforms-sqlalchemy
# psycopg2-binary
psycopg2-binary --only-binary :all:
